% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.r
\name{predict.viord}
\alias{predict.viord}
\title{Posterior predictive probabilities for a fitted \code{viord} model}
\usage{
\method{predict}{viord}(
  object,
  Xn,
  Zn = NULL,
  Y = NULL,
  X = NULL,
  prior = NULL,
  nMC = 1000,
  type = c("class", "prob"),
  ...
)
}
\arguments{
\item{object}{A fitted model of class \code{"viord"}.}

\item{Xn}{A numeric matrix of new covariate values (one row per observation).}

\item{Zn}{Optional random-effects design matrix for new observations. Required
for \code{VB_prior} fits that used random effects.}

\item{Y}{(PMF only) The ordinal response used in fitting the model.}

\item{X}{(PMF only) The design matrix used in fitting the model.}

\item{prior}{(PMF only) A list containing the prior specification
(\code{mu0}, \code{S0}, and \code{Q0}).}

\item{nMC}{Number of Monte Carlo samples for the PMF predictive distribution
(default \code{1e3}).}

\item{type}{A character string specifying the type of prediction:
\code{"prob"} returns the full matrix of posterior predictive probabilities (for each category),
while \code{"class"} returns only the most probable categories
(default \code{"class"}).}

\item{...}{Further arguments (currently ignored).}
}
\value{
Depending on \code{type}:
\itemize{
  \item If \code{type = "prob"}: a numeric matrix of posterior predictive probabilities.
  \item If \code{type = "class"}: a numeric vector of predicted category indices.
}
}
\description{
Computes posterior predictive probabilities or predicted classes for new
observations from a fitted variational ordinal regression model.

Depending on the inference algorithm used to fit the model:
\itemize{
  \item For EP, MF, and VB_prior methods, predictive probabilities are
  obtained analytically since the approximate posterior  is Gaussian
  \item For the PMF method, predictive probabilities are computed
  using a Monte Carlo procedure based on truncated normal sampling.
}
}
\examples{
\dontrun{
fit_mf <- viord(Y, X, prior, method = "MF")
pred_class <- predict(fit_mf, Xn = X_new, type = "class")
pred_prob  <- predict(fit_mf, Xn = X_new, type = "prob")

fit_pmf <- viord(Y, X, prior, method = "PMF")
pred_pmf <- predict(fit_pmf, Y = Y, X = X, prior = prior, Xn = X_new, nMC = 1000)
}

}
