% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viord.r
\name{viord}
\alias{viord}
\title{Variational inference for ordinal models}
\usage{
viord(
  Y,
  X,
  prior,
  algorithm = c("EP", "MF", "PMF"),
  maxit = 100,
  conv_tr = 1e-06
)
}
\arguments{
\item{Y}{Response (ordinal factor)}

\item{X}{Design matrix}

\item{prior}{List with prior parameters (mu0, S0, Q0)}

\item{algorithm}{Character, one of "EP", "MF", or "PMF"}

\item{maxit}{Maximum number of iterations}

\item{conv_tr}{Convergence tolerance}

\item{method}{Optimizer for the thresholds ("grad" or "polr")}
}
\value{
A list containing the estimates and convergence information
}
\description{
Unified interface based for approximate Bayesian inference under the cumulative probit based on Expectation Propagation (EP),
Mean-Field (MF), and Partially Mean-Field (PMF) algorithms.
}
\examples{
\dontrun{
res <- viord(Y, X, prior, algorithm = "MF")
}
}
